analyzer: Separate core and kafka subprojects
