+++ analyzer/core/src/test/scala/io/scalac/newspaper/analyzer/core/AnalyzerSpec.scala
@@ -1,4 +1,4 @@
-package io.scalac.newspaper.analyzer
+package io.scalac.newspaper.analyzer.core
 
 import org.scalatest._
