+++ analyzer/kafka/src/main/resources/application.conf
